feat(ResuambleZoom): add onGestureEnd callback
Creates a callback which is executed and the of a pan gesture and
at the end of a pinch gesture, if the component is out of bounds when
the gesture ends, it will wait for the snapback animation to end.

If decay property is set to true it will wait for the decay animation
to end.