Update BlocksGesture type to allow null references
React states that when a view (like ScrollView, which is a subtype of GestureType) unmounts:

> [React will set the current property back to null when the node is removed from the screen.](https://react.dev/reference/react/useRef#:~:text=React%20will%20set%20the%20current%20property%20back%20to%20null%20when%20the%20node%20is%20removed%20from%20the%20screen.)

This means that in reality, all refs can become null, so `RefObject<GestureType | null>` is a more accurate reflection of the type, especially if the ref is initialized to null using `useRef<GestureType>(null)`

Currently, the typings for BlocksGesture only support undefined refs. This commit adds support for that typing, but leaves the undefined typing for backwards compatibility.